Combining Beacon/AnalyticsBase
