+++ src/main/java/gov/nci/WebAnalytics/Beacon.java
@@ -8,10 +8,25 @@ import java.net.URI;
 import org.apache.http.client.utils.URLEncodedUtils;
 import org.apache.http.message.BasicNameValuePair;
 import org.apache.http.NameValuePair;
+import org.openqa.selenium.WebDriver;
+import net.lightbody.bmp.BrowserMobProxy;
+import net.lightbody.bmp.core.har.Har;
+import net.lightbody.bmp.core.har.HarEntry;
 
 
 public class Beacon {
+	
+	// Constants
+	public static final String S_CODE_NAME = "s_code.js";
+	public static final String S_ACCOUNT = "s_account";
+	public static final String NCI_FUNCTIONS_NAME = "NCIAnalytics";
+	public static final String PAGE_NAME = "www.cancer.gov";
+	public static final String TRACKING_SERVER = "nci.122.2o7.net";
 
+	// Driver object
+	public WebDriver driver;	
+
+	// Beacon properties
 	public URI uri;
 	public String[] suites;	
 	public List<NameValuePair> params;
@@ -217,5 +232,45 @@ public class Beacon {
 		}
 		return rtn;
 	}	
+	
+	
+	
+	/**
+	 * Configure BrowserMob Proxy for Selenium.<br>
+	 * Modified from https://github.com/lightbody/browsermob-proxy#using-with-selenium
+	 * @throws RuntimeException
+	 */
+	// TODO: clean this up - also should this have a return value? 
+	// TODO: set up filters 
+	// TODO: remove duplicates
+	public static void setHar(BrowserMobProxy proxy, List<String> harList) throws RuntimeException {		
+
+		// A HAR (HTTP Archive) is a file format that can be used by HTTP monitoring tools to export collected data. 
+		// BrowserMob Proxy allows us to manipulate HTTP requests and responses, capture HTTP content, 
+	    // and export performance data as a HAR file object.
+	    Har har = proxy.getHar();
+	    
+	    List<HarEntry> entries = har.getLog().getEntries();
+    	System.out.println("Total HAR entries: " + entries.size());
+    	
+	    for (HarEntry entry : entries) {
+	    	if(entry.getRequest().getUrl().contains(TRACKING_SERVER))
+			{
+	    		String result = entry.getRequest().getUrl();
+	    		try {
+					//result = URLDecoder.decode(result, "UTF-8");
+					if(result.contains("pageName=" + PAGE_NAME)) {
+						harList.add(result);
+					}
+				} catch (Exception e) {
+					result = "bleah";
+				} 
+				//System.out.println(result);
+			}
+	    }  
+	    
+		System.out.println("BMP proxy setup done");
+	}
+	
 	
 }
